refactor: Rename Avatar to ProfileAvatar, as this component only displays the avatar given a profile
