+++ ui/src/lib/components/profiles/ProfileAvatar.svelte
